+++ scripts/macros/2014/classFeatures/rogue/soulknife/psionicEnergy.js
@@ -1,4 +1,14 @@
+import {genericUtils} from '../../../../../utils.js';
+
+async function updateScales(origItem, newItemData) {
+    let { scaleIdentifier=null } = genericUtils.getValidScaleIdentifier(origItem.actor, newItemData, psionicEnergy.scaleAliases, 'soulknife');
+    if (!scaleIdentifier) return;
+    genericUtils.setProperty(newItemData, 'system.activities.dnd5eactivity000.damage.parts.0.custom.formula', '@scale.soulknife.' + scaleIdentifier);
+}
 export let psionicEnergy = {
     name: 'Psionic Power: Psionic Energy',
-    version: '1.1.0'
+    aliases: ['Psionic Power'],
+    version: '1.1.0',
+    scaleAliases: ['psionic-power', 'die'],
+    early: updateScales
 };