+++ scripts/macros/2024/classFeatures/bard/lore/peerlessSkill.js
@@ -51,10 +51,16 @@ async function checkSkillLate({trigger: {entity: item, config, roll, actor, opti
     if (!bardicInspiration) return;
     await genericUtils.update(bardicInspiration, {'system.uses.spent': bardicInspiration.system.uses.spent + 1});
 }
+async function updateScales(origItem, newItemData) {
+    let { scaleIdentifier=null } = genericUtils.getValidScaleIdentifier(origItem.actor, newItemData, bardicInspiration.scaleAliases, 'bard');
+    if (!scaleIdentifier) return;
+    genericUtils.setProperty(newItemData, 'flags.chris-premades.config.scaleIdentifier', scaleIdentifier);
+}
 export let peerlessSkill = {
     name: 'Peerless Skill',
     version: '1.2.10',
     rules: 'modern',
+    early: updateScales,
     midi: {
         actor: [
             {